[refactor] Update typeid import refs
