+++ pkg/go.mod
@@ -21,7 +21,6 @@ require (
 	github.com/spf13/cobra v1.8.0
 	github.com/stretchr/testify v1.9.0
 	go.jetify.com/typeid v1.1.0
-	go.jetpack.io/typeid v1.0.1-0.20240410183543-96a4fd53d1e2
 	golang.org/x/oauth2 v0.19.0
 	google.golang.org/protobuf v1.33.0
 )
